+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The version numbers are structured like `GENERATION.MAJOR.MINOR.PATCH`:
 - Add command `heptaconnect:identity-direction:add` in service definition `Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\IdentityDirection\AddDirectionalIdentity` to add an identity direction
 - Add command `heptaconnect:identity-direction:remove` in service definition `Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\IdentityDirection\RemoveDirectionalIdentity` to remove an identity direction
 - Add command `heptaconnect:identity-direction:list` in service definition `Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\IdentityDirection\ListDirectionalIdentities` to list identity directions
+- Add identity directions into evaluation of `\Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\MappingNode\ListMappingNodeSiblings`
 
 ### Changed